Reproducibility.in.Plant.Pathology: A Systematic Quantitative Review of Articles Providing a Basis for Identifying Reproducible Research Efforts the Field of Plant Pathology

A research compendium for the letter to the editor, "Reproducibility in plant pathology: where do we stand and a way forward", published in "xxxxxx".

See also

Author

Maintainer: Adam H. Sparks adamhsparks@gmail.com (ORCID)

Authors:

Emerson M. Del Ponte edelponte@gmail.com (ORCID)

Kaique S. Alves kaiquedsalves@gmail.com (ORCID)

Zachary S. L. Foster zacharyfoster1989@gmail.com (ORCID)

Niklaus J. Grünwald grunwaln@science.oregonstate.edu (ORCID)