Convert doi to bibtex

Convert a digital object identifier (doi) string into a bibtex entry using the web service <http://www.doi2bib.org>.

doi2bib(doi)

# S4 method for character
doi2bib(doi)

Arguments

doi
The character string of the doi.

Value

a bibtex entry as a character string.

Note

Original script from William Keith Morris, <https://gist.github.com/wkmor1/ae2cc96acad472d49c35>. Documented for use in an R package by Adam H. Sparks

Examples

doi2bib("10.4454/jpp.fa.2012.002")

#> [1] "<!DOCTYPE html>\n<html lang=\"en\">\n<head>\n<meta charset=\"utf-8\">\n<title>Error</title>\n</head>\n<body>\n<pre>Cannot GET /doi2bib</pre>\n</body>\n</html>\n"