Import scoring notes checking inter-rater repeatability

Imports, formats data into proper types and calculates the final reproducibility score.

import_interrater_scores()

Value

A [tibble::tibble()] object of reproducibility score notes formatted for use in analysis and manuscript preparation

Examples

notes <- import_notes()
pander::pander(head(notes))
#> 
#> ----------------------------------------------------------------------------
#>            journal                  abbreviation       year   contains_page 
#> ------------------------------ ---------------------- ------ ---------------
#>  Australasian Plant Pathology   Australas Plant Path   2014        29       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2013        51       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2020         8       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2017        75       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2017        16       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2018        11       
#> ----------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> -------------------------------------------------
#>  assignee              doi               doi_url 
#> ---------- ---------------------------- ---------
#>    Adam     10.1007/s13313-013-0228-9    #NAME?  
#> 
#>    Adam     10.1007/s13313-012-0178-7    #NAME?  
#> 
#>  Emerson    10.1007/s13313-019-00670-y   #NAME?  
#> 
#>   Kaique    10.1007/s13313-017-0465-4    #NAME?  
#> 
#>   Kaique    10.1007/s13313-016-0461-0    #NAME?  
#> 
#>   Kaique    10.1007/s13313-017-0532-x    #NAME?  
#> -------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ------------------------------------------------------------------------------
#>             comments              open   repro_inst    art_class    molecular 
#> -------------------------------- ------ ------------ ------------- -----------
#>                NA                 BOTH       2          Applied         0     
#> 
#>  When the article was published   BOTH       2          Applied         0     
#>   I'm not sure that MaxEnt was                                                
#>    freely available, however,                                                 
#>  very recently (1st half 2017)                                                
#>   it's become freely available                                                
#> 
#>                NA                 BOTH       2        Fundamental       0     
#> 
#>                NA                 BOTH       2          Applied         0     
#> 
#>   Original random page number     BOTH       2        Fundamental       1     
#>  not applicable, selected first                                               
#>    following plant pathology                                                  
#>             article                                                           
#> 
#>                NA                 BOTH       2          Applied         1     
#> ------------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ----------------------------------------------------------------
#>  software_cite   software_avail   comp_mthds_avail   data_avail 
#> --------------- ---------------- ------------------ ------------
#>        3               1                 0               0      
#> 
#>        1               3                 0               3      
#> 
#>        1               2                 0               0      
#> 
#>        0               0                 0               0      
#> 
#>        1               2                 0               3      
#> 
#>        2               2                 0               3      
#> ----------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ------------------------------------------------------
#>          software_used            IF_2year   IF_5year 
#> -------------------------------- ---------- ----------
#>             GenStat                1.599      1.708   
#> 
#>              MaxEnt                1.599      1.708   
#> 
#>              MSTATC                1.599      1.708   
#> 
#>                NA                  1.599      1.708   
#> 
#>  iPhyClassifier, BioEdit, MEGA     1.599      1.708   
#> 
#>  Sequence Scanner, MEGA, MAFFT,    1.599      1.708   
#>    MrModeltest, MrBayes, PAUP                         
#> ------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> -----------------------------------
#>    Source: InCites Journal     uid 
#>  Citation Reports, Clarivate       
#>        Analytics, 2022             
#> ----------------------------- -----
#>              NA                 1  
#> 
#>              NA                 2  
#> 
#>              NA                 3  
#> 
#>              NA                 4  
#> 
#>              NA                 5  
#> 
#>              NA                 6  
#> -----------------------------------
#>