Import reproducibility score notes

Imports, formats data into proper types and calculates the final reproducibility score. Also adds the 2 and 5 year impact factor values from InCites Journal Citation Reports for 2018.

import_notes()

Value

A [tibble::tibble()] object of reproducibility score notes formatted for use in analysis and manuscript preparation

Examples

notes <- import_notes()
pander::pander(head(notes))
#> 
#> --------------------------------------------------------------------
#>      journal        abbreviation    year   contains_page   assignee 
#> ------------------ --------------- ------ --------------- ----------
#>  Forest Pathology   Forest Pathol   2012        142          Adam   
#> 
#>  Forest Pathology   Forest Pathol   2013        144          Adam   
#> 
#>  Forest Pathology   Forest Pathol   2021        83           Adam   
#> 
#>  Forest Pathology   Forest Pathol   2014        125          Adam   
#> 
#>  Forest Pathology   Forest Pathol   2015        114          Adam   
#> 
#>  Forest Pathology   Forest Pathol   2018        62         Emerson  
#> --------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ---------------------------------------------------------------------------
#>                doi                  doi_url             comments           
#> ---------------------------------- --------- ------------------------------
#>  10.1111/j.1439-0329.2011.00729.x   #NAME?                 NA              
#> 
#>         10.1111/efp.12011           #NAME?                 NA              
#> 
#>         10.1111/efp.12647           #NAME?      No page numbers in this    
#>                                                journal, so selected first  
#>                                               article that was research in 
#>                                                           2021             
#> 
#>         10.1111/efp.12076           #NAME?                 NA              
#> 
#>         10.1111/efp.12134           #NAME?                 NA              
#> 
#>         10.1111/efp.12376           #NAME?                 NA              
#> ---------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ------------------------------------------------------------------------------
#>  open   repro_inst    art_class    molecular   software_cite   software_avail 
#> ------ ------------ ------------- ----------- --------------- ----------------
#>  BOTH       0          Applied         0             0               0        
#> 
#>  BOTH       0        Fundamental       1             0               0        
#> 
#>  BOTH       0        Fundamental       1             2               1        
#> 
#>  BOTH       0          Applied         1             3               2        
#> 
#>  BOTH       0          Applied         0             3               2        
#> 
#>  BOTH       0          Applied         0             0               0        
#> ------------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> --------------------------------------------------------------------------
#>  comp_mthds_avail   data_avail           software_used           IF_2year 
#> ------------------ ------------ ------------------------------- ----------
#>         0               0                     NA                  1.363   
#> 
#>         0               2                    NSITE                1.363   
#> 
#>         0               1        SPSS, Primer Premier, geNorm,    1.363   
#>                                            RefFinder                      
#> 
#>         0               0           MAFFT, MEGA, PAUP, SAS        1.363   
#> 
#>         0               0                 SAS, Excel              1.363   
#> 
#>         0               0                     NA                  1.363   
#> --------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ----------------------------------------------
#>  IF_5year     Source: InCites Journal     uid 
#>             Citation Reports, Clarivate       
#>                   Analytics, 2022             
#> ---------- ----------------------------- -----
#>   1.595                 NA                 1  
#> 
#>   1.595                 NA                 2  
#> 
#>   1.595                 NA                 3  
#> 
#>   1.595                 NA                 4  
#> 
#>   1.595                 NA                 5  
#> 
#>   1.595                 NA                 6  
#> ----------------------------------------------
#>