Convert doi to bibtex

Convert a digital object identifier (doi) string into a bibtex entry using the web service <http://www.doi2bib.org>.

doi2bib(doi)

# S4 method for character
doi2bib(doi)

Arguments

doi
The character string of the doi.

Value

a bibtex entry as a character string.

Note

Original script from William Keith Morris, <https://gist.github.com/wkmor1/ae2cc96acad472d49c35>. Documented for use in an R package by Adam H Sparks

Examples

doi2bib("10.4454/jpp.fa.2012.002")

#> No encoding supplied: defaulting to UTF-8.

#> [1] " @article{F. Suárez-Estrella_M.A. Bustamante_R. Moral_M.C. Vargas-García_M.J. López_J. Moreno_2012, title={IN VITRO CONTROL OF FUSARIUM WILT USING AGROINDUSTRIAL SUBPRODUCT-BASED COMPOSTS}, volume={94}, ISSN={1125-4653}, url={http://doi.org/10.4454/jpp.fa.2012.002}, DOI={10.4454/jpp.fa.2012.002}, number={1}, journal={Journal of Plant Pathology}, author={F. Suárez-Estrella and M.A. Bustamante and R. Moral and M.C. Vargas-García and M.J. López and J. Moreno}, year={2012} }\n"