Import reproducibility score notes

Imports, formats data into proper types and calculates the final reproducibility score. Also adds the 2 and 5 year impact factor values from InCites Journal Citation Reports for 2018.

import_notes()

Value

A `tibble` object of reproducibility score notes formatted for use in analysis and manuscript preparation

Examples

notes <- import_notes()
pander::pander(head(notes))

#> 
#> ----------------------------------------------------------------------------
#>            journal                  abbreviation       year   contains_page 
#> ------------------------------ ---------------------- ------ ---------------
#>  Australasian Plant Pathology   Australas Plant Path   2013        51       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2014        29       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2018        11       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2017        16       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2017        75       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2017        40       
#> ----------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ------------------------------------------------------------------------------
#>  assignee              doi                         comments              open 
#> ---------- --------------------------- -------------------------------- ------
#>    Adam     10.1007/s13313-012-0178-7   When the article was published   BOTH 
#>                                          I'm not sure that MaxEnt was         
#>                                           freely available, however,          
#>                                         very recently (1st half 2017)         
#>                                          it's become freely available         
#> 
#>    Adam     10.1007/s13313-013-0228-9                 NA                 BOTH 
#> 
#>   Kaique    10.1007/s13313-017-0532-x                 NA                 BOTH 
#> 
#>   Kaique    10.1007/s13313-016-0461-0    Original random page number     TRUE 
#>                                         not applicable, selected first        
#>                                           following plant pathology           
#>                                                    article                    
#> 
#>   Kaique    10.1007/s13313-017-0465-4                 NA                 BOTH 
#> 
#>   Kaique    10.1007/s13313-016-0458-8                 NA                 BOTH 
#> ------------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> -----------------------------------------------------------------------
#>  repro_inst    art_class    molecular   software_cite   software_avail 
#> ------------ ------------- ----------- --------------- ----------------
#>      2          Applied       FALSE           3               3        
#> 
#>      2          Applied       FALSE           3               2        
#> 
#>      2          Applied       TRUE            2               3        
#> 
#>      2        Fundamental     TRUE            1               3        
#> 
#>      2          Applied       FALSE           0               0        
#> 
#>      2        Fundamental     TRUE            1               3        
#> -----------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ---------------------------------------------------------------------------
#>  comp_mthds_avail   data_avail           software_used            category 
#> ------------------ ------------ -------------------------------- ----------
#>         0               2                    MaxEnt               ARTICLE  
#> 
#>         0               0                   GenStat                  NA    
#> 
#>         0               3        Sequence Scanner, MEGA, MAFFT,      NA    
#>                                    MrModeltest, MrBayes, PAUP              
#> 
#>         0               3        iPhyClassifier, BioEdit, MEGA       NA    
#> 
#>         0               0                      NA                    NA    
#> 
#>         0               0               POPGENE, GenAlEx             NA    
#> ---------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> --------------------------------------------------------------------------
#>  bibtexkey   address   annote              author               booktitle 
#> ----------- --------- -------- ------------------------------- -----------
#>     NA         NA        NA     J Elith, J Simpson, M Hirsch,      NA     
#>                                          M A Burgman                      
#> 
#>     NA         NA        NA      R. Billones-Baaijens, E. E.       NA     
#>                                  Jones, H. J. Ridgway, M. V.              
#>                                            Jaspers                        
#> 
#>     NA         NA        NA                 NULL                   NA     
#> 
#>     NA         NA        NA                 NULL                   NA     
#> 
#>     NA         NA        NA                 NULL                   NA     
#> 
#>     NA         NA        NA                 NULL                   NA     
#> --------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> --------------------------------------------------------------------------
#>  chapter   crossref   edition   editor   howpublished   institution   key 
#> --------- ---------- --------- -------- -------------- ------------- -----
#>    NA         NA        NA       NULL         NA            NA        NA  
#> 
#>    NA         NA        NA       NULL         NA            NA        NA  
#> 
#>    NA         NA        NA        NA          NA            NA        NA  
#> 
#>    NA         NA        NA        NA          NA            NA        NA  
#> 
#>    NA         NA        NA        NA          NA            NA        NA  
#> 
#>    NA         NA        NA        NA          NA            NA        NA  
#> --------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> -----------------------------------------------------------------------------
#>  month   note   number   organization   pages    publisher   school   series 
#> ------- ------ -------- -------------- -------- ----------- -------- --------
#>   jan     NA      1           NA        43--51      NA         NA       NA   
#> 
#>   may     NA      1           NA        25--31      NA         NA       NA   
#> 
#>   NA      NA      NA          NA          NA        NA         NA       NA   
#> 
#>   NA      NA      NA          NA          NA        NA         NA       NA   
#> 
#>   NA      NA      NA          NA          NA        NA         NA       NA   
#> 
#>   NA      NA      NA          NA          NA        NA         NA       NA   
#> -----------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ------------------------------------------------------------
#>              title               type   volume   url   issn 
#> ------------------------------- ------ -------- ----- ------
#>    Taxonomic uncertainty and      NA      42     NA     NA  
#>       decision making for                                   
#>   biosecurity: spatial models                               
#>      for myrtle/guava rust                                  
#> 
#>    Susceptibility of common       NA      43     NA     NA  
#>  rootstock and scion varieties                              
#>        of grapevines to                                     
#>  \textit{{Botryosphaeriaceae}}                              
#>             species                                         
#> 
#>               NA                  NA      NA     NA     NA  
#> 
#>               NA                  NA      NA     NA     NA  
#> 
#>               NA                  NA      NA     NA     NA  
#> 
#>               NA                  NA      NA     NA     NA  
#> ------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> -------------------------------------------------------------------------
#>  x.article.pm..title   IF_2year   IF_5year     Source: InCites Journal   
#>                                              Citation Reports, Clarivate 
#>                                                    Analytics, 2018       
#> --------------------- ---------- ---------- -----------------------------
#>          NA             1.106       1.36                 NA              
#> 
#>          NA             1.106       1.36                 NA              
#> 
#>          NA             1.106       1.36                 NA              
#> 
#>          NA             1.106       1.36                 NA              
#> 
#>          NA             1.106       1.36                 NA              
#> 
#>          NA             1.106       1.36                 NA              
#> -------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ----------------------------------------
#>  total_possible   reproducibility_score 
#> ---------------- -----------------------
#>        12                 66.67         
#> 
#>        12                 41.67         
#> 
#>        12                 66.67         
#> 
#>        12                 58.33         
#> 
#>        12                   0           
#> 
#>        12                 33.33         
#> ----------------------------------------
#>