Import Reproducibility Score Notes

Imports, formats data into proper types and calculates the final reproducibility score. Also adds the 2 and 5 year impact factor values from InCites Journal Citation Reports for 2018.

import_notes()

Value

A `tibble` object of reproducibility score notes formatted for use in analysis and manuscript preparation

Examples

notes <- import_notes()

#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   .default = col_character(),
#>   year = col_double(),
#>   contains_page = col_double(),
#>   repro_inst = col_double(),
#>   molecular = col_logical(),
#>   software_cite = col_double(),
#>   software_avail = col_double(),
#>   comp_mthds_avail = col_double(),
#>   data_avail = col_double(),
#>   address = col_logical(),
#>   annote = col_logical(),
#>   booktitle = col_logical(),
#>   chapter = col_logical(),
#>   crossref = col_logical(),
#>   edition = col_logical(),
#>   howpublished = col_logical(),
#>   institution = col_logical(),
#>   key = col_logical(),
#>   note = col_logical(),
#>   number = col_double(),
#>   organization = col_logical()
#>   # ... with 5 more columns
#> )
#> ℹ Use `spec()` for the full column specifications.

#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   journal = col_character(),
#>   IF_2year = col_double(),
#>   IF_5year = col_double(),
#>   `Source: InCites Journal Citation Reports, Clarivate Analytics, 2018` = col_logical()
#> )

pander::pander(head(notes))

#> 
#> ----------------------------------------------------------------------------
#>            journal                  abbreviation       year   contains_page 
#> ------------------------------ ---------------------- ------ ---------------
#>  Australasian Plant Pathology   Australas Plant Path   2013        13       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2013        51       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2014        29       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2014        55       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2017        75       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2018        11       
#> ----------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ------------------------------------------------------------------------------
#>  assignee              doi                         comments              open 
#> ---------- --------------------------- -------------------------------- ------
#>    Zach     10.1007/s13313-012-0157-z   The journal has guidelines for   BOTH 
#>                                          data availability, but does          
#>                                          not mention reproducibility          
#>                                           exactly. I am not sure how          
#>                                          this effects the "repo_inst"         
#>                                          value. For the "country", I          
#>                                          put the country of where the         
#>                                                authors worked.                
#> 
#>    Adam     10.1007/s13313-012-0178-7   When the article was published   BOTH 
#>                                          I'm not sure that MaxEnt was         
#>                                           freely available, however,          
#>                                         very recently (1st half 2017)         
#>                                          it's become freely available         
#> 
#>    Adam     10.1007/s13313-013-0228-9                 NA                 BOTH 
#> 
#>    Zach     10.1007/s13313-013-0240-0   The journal has guidelines for   BOTH 
#>                                          data availability, but does          
#>                                          not mention reproducibility          
#>                                           exactly. I am not sure how          
#>                                          this effects the "repo_inst"         
#>                                          value. For the "country", I          
#>                                          put the country of where the         
#>                                                authors worked.                
#> 
#>   Kaique    10.1007/s13313-017-0465-4                 NA                 BOTH 
#> 
#>   Kaique    10.1007/s13313-017-0532-x                 NA                 BOTH 
#> ------------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ---------------------------------------------------------------------
#>  repro_inst   art_class   molecular   software_cite   software_avail 
#> ------------ ----------- ----------- --------------- ----------------
#>      2         Applied      FALSE           0               0        
#> 
#>      2         Applied      FALSE           3               3        
#> 
#>      2         Applied      FALSE           3               2        
#> 
#>      2         Applied      FALSE           0               0        
#> 
#>      2         Applied      FALSE           0               0        
#> 
#>      2         Applied      TRUE            2               3        
#> ---------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ---------------------------------------------------------------------------
#>  comp_mthds_avail   data_avail           software_used            category 
#> ------------------ ------------ -------------------------------- ----------
#>         0               0                      NA                    NA    
#> 
#>         0               2                    MaxEnt               ARTICLE  
#> 
#>         0               0                   GenStat                  NA    
#> 
#>         0               0                      NA                    NA    
#> 
#>         0               0                      NA                    NA    
#> 
#>         0               3        Sequence Scanner, MEGA, MAFFT,      NA    
#>                                    MrModeltest, MrBayes, PAUP              
#> ---------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> --------------------------------------------------------------------------
#>  bibtexkey   address   annote              author               booktitle 
#> ----------- --------- -------- ------------------------------- -----------
#>     NA         NA        NA                 NULL                   NA     
#> 
#>     NA         NA        NA     J Elith, J Simpson, M Hirsch,      NA     
#>                                          M A Burgman                      
#> 
#>     NA         NA        NA      R. Billones-Baaijens, E. E.       NA     
#>                                  Jones, H. J. Ridgway, M. V.              
#>                                            Jaspers                        
#> 
#>     NA         NA        NA                 NULL                   NA     
#> 
#>     NA         NA        NA                 NULL                   NA     
#> 
#>     NA         NA        NA                 NULL                   NA     
#> --------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> --------------------------------------------------------------------------
#>  chapter   crossref   edition   editor   howpublished   institution   key 
#> --------- ---------- --------- -------- -------------- ------------- -----
#>    NA         NA        NA       NULL         NA            NA        NA  
#> 
#>    NA         NA        NA       NULL         NA            NA        NA  
#> 
#>    NA         NA        NA       NULL         NA            NA        NA  
#> 
#>    NA         NA        NA       NULL         NA            NA        NA  
#> 
#>    NA         NA        NA        NA          NA            NA        NA  
#> 
#>    NA         NA        NA        NA          NA            NA        NA  
#> --------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> -----------------------------------------------------------------------------
#>  month   note   number   organization   pages    publisher   school   series 
#> ------- ------ -------- -------------- -------- ----------- -------- --------
#>   jan     NA      1           NA          NA        NA         NA       NA   
#> 
#>   jan     NA      1           NA        43--51      NA         NA       NA   
#> 
#>   may     NA      1           NA        25--31      NA         NA       NA   
#> 
#>   jul     NA      1           NA        53--60      NA         NA       NA   
#> 
#>   NA      NA      NA          NA          NA        NA         NA       NA   
#> 
#>   NA      NA      NA          NA          NA        NA         NA       NA   
#> -----------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ------------------------------------------------------------
#>              title               type   volume   url   issn 
#> ------------------------------- ------ -------- ----- ------
#>  On–farm waste-based composts     NA      42     NA     NA  
#>           in managing                                       
#>      \textit{{Macrophomina                                  
#>  phaseolina}} induced dry root                              
#>     rot of guar in an arid                                  
#>           environment                                       
#> 
#>    Taxonomic uncertainty and      NA      42     NA     NA  
#>       decision making for                                   
#>   biosecurity: spatial models                               
#>      for myrtle/guava rust                                  
#> 
#>    Susceptibility of common       NA      43     NA     NA  
#>  rootstock and scion varieties                              
#>        of grapevines to                                     
#>  \textit{{Botryosphaeriaceae}}                              
#>             species                                         
#> 
#>  An improved culture solution     NA      43     NA     NA  
#>          technique for                                      
#>     \textit{{Plasmodiophora                                 
#>  brassicae}} infection and the                              
#>  dynamic infection in the root                              
#>              hair                                           
#> 
#>               NA                  NA      NA     NA     NA  
#> 
#>               NA                  NA      NA     NA     NA  
#> ------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> -------------------------------------------------------------------------
#>  x.article.pm..title   IF_2year   IF_5year     Source: InCites Journal   
#>                                              Citation Reports, Clarivate 
#>                                                    Analytics, 2018       
#> --------------------- ---------- ---------- -----------------------------
#>          NA             1.106       1.36                 NA              
#> 
#>          NA             1.106       1.36                 NA              
#> 
#>          NA             1.106       1.36                 NA              
#> 
#>          NA             1.106       1.36                 NA              
#> 
#>          NA             1.106       1.36                 NA              
#> 
#>          NA             1.106       1.36                 NA              
#> -------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ----------------------------------------
#>  total_possible   reproducibility_score 
#> ---------------- -----------------------
#>        12                   0           
#> 
#>        12                 66.67         
#> 
#>        12                 41.67         
#> 
#>        12                   0           
#> 
#>        12                   0           
#> 
#>        12                 66.67         
#> ----------------------------------------
#>