Import reproducibility score notes

Imports, formats data into proper types and calculates the final reproducibility score. Also adds the 2 and 5 year impact factor values from InCites Journal Citation Reports for 2018.

import_notes()

Value

A [tibble::tibble()] object of reproducibility score notes formatted for use in analysis and manuscript preparation

Examples

notes <- import_notes()
pander::pander(head(notes))
#> 
#> ----------------------------------------------------------------------------
#>            journal                  abbreviation       year   contains_page 
#> ------------------------------ ---------------------- ------ ---------------
#>  Australasian Plant Pathology   Australas Plant Path   2013        51       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2013        13       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2014        29       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2014        55       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2017        16       
#> 
#>  Australasian Plant Pathology   Australas Plant Path   2017        75       
#> ----------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ------------------------------------------------
#>  assignee              doi              doi_url 
#> ---------- --------------------------- ---------
#>    Adam     10.1007/s13313-012-0178-7   #NAME?  
#> 
#>    Zach     10.1007/s13313-012-0157-z   #NAME?  
#> 
#>    Adam     10.1007/s13313-013-0228-9   #NAME?  
#> 
#>    Zach     10.1007/s13313-013-0240-0   #NAME?  
#> 
#>   Kaique    10.1007/s13313-016-0461-0   #NAME?  
#> 
#>   Kaique    10.1007/s13313-017-0465-4   #NAME?  
#> ------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ------------------------------------------------------------------------------
#>             comments              open   repro_inst    art_class    molecular 
#> -------------------------------- ------ ------------ ------------- -----------
#>  When the article was published   BOTH       2          Applied         0     
#>   I'm not sure that MaxEnt was                                                
#>    freely available, however,                                                 
#>  very recently (1st half 2017)                                                
#>   it's become freely available                                                
#> 
#>  The journal has guidelines for   BOTH       2          Applied         0     
#>   data availability, but does                                                 
#>   not mention reproducibility                                                 
#>    exactly. I am not sure how                                                 
#>   this effects the "repo_inst"                                                
#>   value. For the "country", I                                                 
#>   put the country of where the                                                
#>         authors worked.                                                       
#> 
#>                NA                 BOTH       2          Applied         0     
#> 
#>  The journal has guidelines for   BOTH       2          Applied         0     
#>   data availability, but does                                                 
#>   not mention reproducibility                                                 
#>    exactly. I am not sure how                                                 
#>   this effects the "repo_inst"                                                
#>   value. For the "country", I                                                 
#>   put the country of where the                                                
#>         authors worked.                                                       
#> 
#>   Original random page number     BOTH       2        Fundamental       1     
#>  not applicable, selected first                                               
#>    following plant pathology                                                  
#>             article                                                           
#> 
#>                NA                 BOTH       2          Applied         0     
#> ------------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ----------------------------------------------------------------
#>  software_cite   software_avail   comp_mthds_avail   data_avail 
#> --------------- ---------------- ------------------ ------------
#>        3               3                 0               2      
#> 
#>        0               0                 0               0      
#> 
#>        3               2                 0               0      
#> 
#>        0               0                 0               0      
#> 
#>        1               3                 0               3      
#> 
#>        0               0                 0               0      
#> ----------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> -------------------------------------------------------------------------
#>          software_used           category   bibtexkey   address   annote 
#> ------------------------------- ---------- ----------- --------- --------
#>             MaxEnt               ARTICLE       NA         NA        NA   
#> 
#>               NA                    NA         NA         NA        NA   
#> 
#>             GenStat                 NA         NA         NA        NA   
#> 
#>               NA                    NA         NA         NA        NA   
#> 
#>  iPhyClassifier, BioEdit, MEGA      NA         NA         NA        NA   
#> 
#>               NA                    NA         NA         NA        NA   
#> -------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> --------------------------------------------------------------------------
#>             author               booktitle   chapter   crossref   edition 
#> ------------------------------- ----------- --------- ---------- ---------
#>  J Elith, J Simpson, M Hirsch,      NA         NA         NA        NA    
#>           M A Burgman                                                     
#> 
#>              NULL                   NA         NA         NA        NA    
#> 
#>   R. Billones-Baaijens, E. E.       NA         NA         NA        NA    
#>   Jones, H. J. Ridgway, M. V.                                             
#>             Jaspers                                                       
#> 
#>              NULL                   NA         NA         NA        NA    
#> 
#>              NULL                   NA         NA         NA        NA    
#> 
#>              NULL                   NA         NA         NA        NA    
#> --------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> -------------------------------------------------------------------
#>  editor   howpublished   institution   key   month   note   number 
#> -------- -------------- ------------- ----- ------- ------ --------
#>   NULL         NA            NA        NA     jan     NA      1    
#> 
#>   NULL         NA            NA        NA     jan     NA      1    
#> 
#>   NULL         NA            NA        NA     may     NA      1    
#> 
#>   NULL         NA            NA        NA     jul     NA      1    
#> 
#>    NA          NA            NA        NA     NA      NA      NA   
#> 
#>    NA          NA            NA        NA     NA      NA      NA   
#> -------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> -----------------------------------------------------
#>  organization   pages    publisher   school   series 
#> -------------- -------- ----------- -------- --------
#>       NA        43--51      NA         NA       NA   
#> 
#>       NA          NA        NA         NA       NA   
#> 
#>       NA        25--31      NA         NA       NA   
#> 
#>       NA        53--60      NA         NA       NA   
#> 
#>       NA          NA        NA         NA       NA   
#> 
#>       NA          NA        NA         NA       NA   
#> -----------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ------------------------------------------------------------
#>              title               type   volume   url   issn 
#> ------------------------------- ------ -------- ----- ------
#>    Taxonomic uncertainty and      NA      42     NA     NA  
#>       decision making for                                   
#>   biosecurity: spatial models                               
#>      for myrtle/guava rust                                  
#> 
#>  On–farm waste-based composts     NA      42     NA     NA  
#>           in managing                                       
#>      \textit{{Macrophomina                                  
#>  phaseolina}} induced dry root                              
#>     rot of guar in an arid                                  
#>           environment                                       
#> 
#>    Susceptibility of common       NA      43     NA     NA  
#>  rootstock and scion varieties                              
#>        of grapevines to                                     
#>  \textit{{Botryosphaeriaceae}}                              
#>             species                                         
#> 
#>  An improved culture solution     NA      43     NA     NA  
#>          technique for                                      
#>     \textit{{Plasmodiophora                                 
#>  brassicae}} infection and the                              
#>  dynamic infection in the root                              
#>              hair                                           
#> 
#>               NA                  NA      NA     NA     NA  
#> 
#>               NA                  NA      NA     NA     NA  
#> ------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> -------------------------------------------------------------------------
#>  x.article.pm..title   IF_2year   IF_5year     Source: InCites Journal   
#>                                              Citation Reports, Clarivate 
#>                                                    Analytics, 2022       
#> --------------------- ---------- ---------- -----------------------------
#>          NA             1.599      1.708                 NA              
#> 
#>          NA             1.599      1.708                 NA              
#> 
#>          NA             1.599      1.708                 NA              
#> 
#>          NA             1.599      1.708                 NA              
#> 
#>          NA             1.599      1.708                 NA              
#> 
#>          NA             1.599      1.708                 NA              
#> -------------------------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> ----------------------------------------
#>  total_possible   reproducibility_score 
#> ---------------- -----------------------
#>        12                 66.67         
#> 
#>        12                   0           
#> 
#>        12                 41.67         
#> 
#>        12                   0           
#> 
#>        12                 58.33         
#> 
#>        12                   0           
#> ----------------------------------------
#>